requirements.txt
requests
python-dotenv
pandas
## py -m pip install -r requirements.txt